refactor(analysis-types): split topics DTOs
